feat: refactor backup classes
