+++ src/cli/github/backup/issue.ts
@@ -3,20 +3,19 @@ import fs from 'fs-extra';
 import { cpus } from 'os';
 import pAll from 'p-all';
 import { dirname } from 'path';
-import { AbstractGithubBackup } from '../protocol';
+import { AbstractGithubBackup, Repo } from '../protocol';
 
 export class GithubIssueBackup extends AbstractGithubBackup {
 
     async backup() {
-        const { org } = this.options;
         const repos = await this.octokit.repos.listForOrg({
-            org: org
+            org: this.options.org
         });
 
         await pAll(
             repos.data.map(repo => {
                 return async () => {
-                    await this.backupIssues(repo.name);
+                    await this.backupRepository(repo);
                 }
             }),
             {
@@ -25,8 +24,9 @@ export class GithubIssueBackup extends AbstractGithubBackup {
         );
     }
 
-    async backupIssues(repoName: string): Promise<void> {
+    async backupRepository(repo: Repo): Promise<void> {
 
+        const repoName = repo.name;
         const issues: RestEndpointMethodTypes["issues"]["listForRepo"]["response"] = await this.octokit.issues.listForRepo({
             owner: this.options.org,
             repo: repoName,
@@ -44,7 +44,7 @@ export class GithubIssueBackup extends AbstractGithubBackup {
                     })
 
                     const data = { ...issue, extra: { comments } }
-                    const issuePath = this.getIssuePath(repoName, issue.id)
+                    const issuePath = this.getIssuePath(repoName, issue.number)
 
                     await fs.ensureDir(dirname(issuePath))
                     await fs.writeJson(issuePath, data);
