+++ src/cli/github/backup/settings.ts
@@ -3,20 +3,19 @@ import fs from 'fs-extra';
 import { cpus } from 'os';
 import pAll from 'p-all';
 import { dirname } from 'path';
-import { AbstractGithubBackup } from '../protocol';
+import { AbstractGithubBackup, Repo } from '../protocol';
 
 export class GithubSettingsBackup extends AbstractGithubBackup {
 
     async backup() {
-        const { org } = this.options;
         const repos = await this.octokit.repos.listForOrg({
-            org: org
+            org: this.options.org
         });
 
         await pAll(
             repos.data.map(repo => {
                 return async () => {
-                    await this.backupSettings(repo.name);
+                    await this.backupRepository(repo);
                 }
             }),
             {
@@ -25,7 +24,9 @@ export class GithubSettingsBackup extends AbstractGithubBackup {
         );
     }
 
-    async backupSettings(repoName: string): Promise<void> {
+    async backupRepository(repo: Repo): Promise<void> {
+
+        const repoName = repo.name;
         const settings: RestEndpointMethodTypes["repos"]["get"]["response"] = await this.octokit.repos.get({
             owner: this.options.org,
             repo: repoName,
